#  codegen: ensure Bool arithmetic is done in mod-1