As @stevengj found in #31486, we're doing boolean arithmetic wrong in codegen, and that's causing us to potentially give up lots of performance since we can't vectorize it, and in addition we get wrong answers.

setup:
```julia
julia> xor_(a, b) = Core.Intrinsics.xor_int(a, b);
julia> b = rand(Bool, 150); i = UInt8.(b);
julia> f(a) = reduce(xor, a); g(a) = reduce(!=, a); h(a) = reduce(xor_, a);
julia> using BenchmarkTools
```

before:
```
julia> @btime f(b)
  115.304 ns (0 allocations: 0 bytes)
true

julia> @btime f(i)
  34.677 ns (0 allocations: 0 bytes)
0x01

julia> @btime g(b)
  114.314 ns (0 allocations: 0 bytes)
true

julia> @btime g(i)
  119.157 ns (0 allocations: 0 bytes)
true

julia> @btime h(b)
  33.699 ns (0 allocations: 0 bytes)
true

julia> @btime h(i)
  35.412 ns (0 allocations: 0 bytes)
0x01
```

after:
```
julia> @btime f(b)
  38.064 ns (0 allocations: 0 bytes)
true

julia> @btime f(i)
  34.773 ns (0 allocations: 0 bytes)
0x01

julia> @btime g(b)
  35.565 ns (0 allocations: 0 bytes)
true

julia> @btime g(i)
  118.535 ns (0 allocations: 0 bytes)
true

julia> @btime h(b)
  39.737 ns (0 allocations: 0 bytes)
true

julia> @btime h(i)
  34.269 ns (0 allocations: 0 bytes)
0x01

julia> @btime h(i)
  34.299 ns (0 allocations: 0 bytes)
0x01
```

PS: ignore the first commit, it's just deleting dead code