@nanosoldier 2019-04-02
Something went wrong when running [your job](https://github.com/JuliaLang/julia/pull/31503#issuecomment-479102998):
```
NanosoldierError: failed to run benchmarks against primary commit: failed process: Process(`make -j3 USECCACHE=1`, ProcessExited(2)) [2]
```
Logs and partial data can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/6bad8ea0084c8dd0d2a4ff9d9147d20dd59091dc/d9f0608_vs_98673b5/report.md)
cc @ararslan