@ararslan 2019-04-02
Nanosoldier is currently broken due to the BinaryBuilder changes.