@nanosoldier 2019-04-03
[Your benchmark job](https://github.com/JuliaLang/julia/pull/31503#issuecomment-479203435) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/078c632e89f610ed0149cd22adec840ddeee6190/d9f0608_vs_98673b5/report.md). cc @ararslan