@nanosoldier 2019-04-03
[Your benchmark job](https://github.com/JuliaLang/julia/pull/31503#issuecomment-479327971) has completed - possible performance regressions were detected. A full report can be found [here](https://github.com/JuliaCI/BaseBenchmarkReports/blob/1fa11c3d5782d998c183147c6141e64cc1cfb6f0/044d56d_vs_aee211b/report.md). cc @ararslan