@rfourquet 2019-04-03
"mod 2" instead of "mod 1" ?